refactor(core): reorganize swarm core into a module
